% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SerialNumber.R
\name{SerialNumberUI}
\alias{SerialNumberUI}
\title{序列号查询生成器界面}
\usage{
SerialNumberUI(
  tabTitle = "Serial Number",
  colTitles = c("Operation Area", "Column Setting", "Result Area"),
  widthRates = c(6, 6, 12),
  func_left = SerialNumberUI_left,
  func_right = SerialNumberUI_right,
  func_bottom = SerialNumberUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
序列号查询生成器界面
}
\examples{
SerialNumberUI()
}
